Global client
Both `cache` and `rateLimit` options are worthless if the client is recreated for each lambda execution.

Due to the stateless natura of lambda functions, this is not able to ensure proper caching, but at least on hot starts the client will be reused for each lambda container which is better than nothing.

Am I missing anything? Is the client cache somehow working even when recreated each time?